refactor(health): move logic to rpc submodule, fix dep ordering
