+++ crates/proof/tee/nitro/Cargo.toml
@@ -15,11 +15,11 @@ workspace = true
 # Core
 base-alloy-evm.workspace = true
 base-proof-client.workspace = true
+tower = { workspace = true, optional = true }
 base-health = { workspace = true, optional = true }
 base-proof-host = { workspace = true, optional = true }
 base-proof-preimage = { workspace = true, features = ["serde"] }
 base-proof-primitives = { workspace = true, features = ["std", "serde"] }
-tower = { workspace = true, optional = true }
 jsonrpsee = { workspace = true, optional = true, features = ["server", "macros"] }
 
 # Alloy
